requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
Django==4.1.2
iniconfig==1.1.1
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
qrcode==7.3.1
sqlparse==0.4.3
tomli==2.0.1
